requirements.txt
feedparser==6.0.10
requests==2.31.0
openai==1.12.0
jinja2==3.1.3
pyyaml==6.0.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
schedule==1.2.1
